move RunicArgs into Rune.ts
