+++ rune/mod.ts
@@ -5,6 +5,5 @@ export * from "./FnRune.ts"
 export * from "./Id.ts"
 export * from "./MetaRune.ts"
 export * from "./Rune.ts"
-export * from "./RunicArgs.ts"
 export * from "./Timeline.ts"
 export * from "./ValueRune.ts"
